Disabled controllers and created sample endpoint to ensure prisma is querying correctly
