+++ api/package.json
@@ -3,7 +3,7 @@
   "version": "1.0.0",
   "main": "index.js",
   "scripts": {
-    "start": "nodemon index.js",
+    "start": "ts-node-dev index.ts",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "",
@@ -11,13 +11,16 @@
   "description": "",
   "dependencies": {
     "@prisma/client": "^6.14.0",
+    "@types/express": "^5.0.3",
     "cors": "^2.8.5",
     "express": "^4.19.2",
     "multer": "^1.4.5-lts.1",
-    "nodemon": "^3.1.4",
     "sqlite3": "^5.1.7"
   },
   "devDependencies": {
-    "prisma": "^6.14.0"
+    "@types/cors": "^2.8.19",
+    "@types/node": "^24.3.0",
+    "prisma": "^6.14.0",
+    "ts-node-dev": "^2.0.0"
   }
 }
